% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LARMC}
\alias{LARMC}
\title{A function to generate Markov chains using Rcpp}
\usage{
LARMC(sigma, x0, N)
}
\arguments{
\item{sigma}{The variance of the normal distribution}

\item{x0}{The initial value}

\item{N}{The length of the chain}
}
\value{
a Markov chain with length \code{N}
}
\description{
A function to generate Markov chains using Rcpp given the variance of the normal distribution, the initial value and the length of the chain.
}
\examples{
\dontrun{
sigma=1
x1=0
N=2000
X=LARMC(sigma,x0,N)
plot(x)
}
}
