% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20087.R
\name{NRFE}
\alias{NRFE}
\title{Non recursive feature elimination (NRFE)}
\usage{
NRFE(dt, ntree)
}
\arguments{
\item{dt}{The initial data frame}

\item{ntree}{The number of decision trees in the forest}
}
\value{
A vector \code{err}
}
\description{
A non recursive function using random forest to select variables and calculate the error.This function can be used to 
calculate the PI for classification and regression problems. Pay attention that in the input data frame, the dependent variable 'Y' should be placed in the second column.
}
\examples{
\dontrun{
dt=longley
ntree=800
NRFE(dt,ntree)
} 
}
