% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20087.R
\name{permudata}
\alias{permudata}
\title{A permutation function using R}
\usage{
permudata(dt, i)
}
\arguments{
\item{dt}{The initial data frame}

\item{i}{The label of variable to be permuted}
}
\value{
A permuted data frame
}
\description{
Change the label of a certain variable to calculate its importance
}
\examples{
\dontrun{
dt=iris
i=3
permudata(dt,i)
} 
}
