% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20087.R
\name{rfpi}
\alias{rfpi}
\title{A function using R to calculate the permutation importance}
\usage{
rfpi(dt, ntree)
}
\arguments{
\item{dt}{The initial data frame}

\item{ntree}{The number of decision trees in the forest}
}
\value{
The permutation importance vector of each variable\code{PI}
}
\description{
Calculate the permutation importance of variables with random forest algorithm. This function can be used to 
calculate the PI for classification and regression problems. Pay attention that in the input data frame, the dependent variable 'Y' should be placed in the second column.
}
\examples{
\dontrun{
dt=longley
ntree=500
rfpi(dt,ntree)
} 
}
