\name{StatComp20087-package}
\alias{StatComp20087-package}
\docType{package}
\title{
  This package shows the homework in the 'Statistical Computing' course and some functions in my personal research. 
}
\description{
  The research functions are used to calculate variable permutation importance in random forests(\code{rfpi}, \code{permudata}). The package provided two algorithms: NRFE(non recursive feature elimination) and RFE(recursive feature elimination)(\code{RFE} and \code{NRFE}).
}

\author{
Jingzhi Ding

Maintainer: Jingzhi Ding <dingjz1998@mail.ustc.edu.cn>
}

\keyword{ package }

